% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damages}
\format{
A data frame with two columns
\describe{
  \item{distance}{how far}
  \item{damage}{how bad}
  ...
}
}
\source{
I made it up
}
\usage{
fire
}
\description{
A dataset that contains how far a fire traveled and how much damage it caused
}
\keyword{datasets}
