% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{how many numbers}

\item{iter}{how many times to iterate}

\item{a}{Default a value}

\item{b}{Default b value}
}
\value{
A bargraph!
}
\description{
Central Limit Theorem Function
}
