% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Title}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{is the frame of data to be passed to the function}

\item{SPECIES}{is the subset of the data, the SPECIES, to analyze}
}
\value{

}
\description{
Title
}
\examples{
myddt(read.csv("/Users/marcosbernier/Dev/stats/MATH4753BERN0021/projects/project1/DDT.csv"),"CCATFISH")
}
