% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Title}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{is the dataset vector}

\item{mn}{is the name of the barplot}

\item{...}{}
}
\value{
Returns a pareto barplot
}
\description{
Title
}
